requirements.txt
pickle5
chalice
matplotlib
sklearn
numpy
scipy
pandas
flask-restful
zeyrek
